requirements.txt
nba_api>=1.11.4
requests>=2.31.0
psycopg2-binary>=2.9.9
gspread>=5.12.0
google-auth>=2.23.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.0